god this took forever
We didn't learn enough about shaders to make our own this project.

But I mean, I still made it.
It took hours to get  the self made shader because I tried to replicate one from the website. But there are no comments in those codes. So I gave up and made this.